\name{NormalRegPanelMixture-package}
\alias{NormalRegPanelMixture-package}
\alias{NormalRegPanelMixture}
\docType{package}
\title{
\packageTitle{NormalRegPanelMixture}
}
\description{
\packageDescription{NormalRegPanelMixture}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NormalRegPanelMixture}
\packageIndices{NormalRegPanelMixture}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NormalRegPanelMixture}

Maintainer: \packageMaintainer{NormalRegPanelMixture}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
